Getting there with type identity and exports
I still have some questions about how and where the type-checker resolves
aliases.  One of the things that happens is two types will be compared for
equality when they are t-names, but one of them is an alias.  The idea of
aliases-of-aliases, as in

type A = B
type C = A
type D = C

needs to be worked out -- should these all be collapsed to "B" in the export
immediately, or should a recursive "resolve-aliases" be present in the
type-checker?

I think that t-name is doing too many jobs right now, and needs to be split
along the lines of brander-defined types, datatypes, and aliases, with a clear
definition of what it means to look up/resolve a type.